Make libfec an internal library
Due to the custom functions required on the libfec, it can not be an
external dependency. This commit, adds the libfec as an internal library
and installs it on the system. To avoid conflicts with an already
installed libfec, the symbol name follows the GNU Radio shared library
naming scheme (libgnuradio-ccsds-fec.so).